support use of when in tasks
This field relies on local time,
and will run a task once during each period when the time specification of when matches the current local time.
Currently, If a task is set to run during a jump forward or backward,
the task will run 0 times during a jump forward and two times during a jump backward.