return exit code without signal/state information attached
Return the exit code we would see in the shell.
Process.exit_status returns a code that combines (a possible) signal status, along with the exit code.
We just want the exit code. If a process is killed by a signal, we can examine system logs if crdo's logs aren't enough.